renamed component files
