+++ client/components/index.js
@@ -3,8 +3,8 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Main} from './main'
-export {default as UserHome} from './user-home'
+export {default as Main} from './Main'
+export {default as UserHome} from './UserHome'
 export {Login, Signup} from './AuthForm'
 export {default as AllWatches} from './AllWatches'
 export {default as AllLessons} from './lessons'
