+++ server/db/models/order.js
@@ -1,0 +1,12 @@
+const Sequelize = require('sequelize')
+const db = require('../db')
+
+const Order = db.define('order', {
+  status: {
+    type: Sequelize.ENUM,
+    values: ['cart', 'created', 'processing', 'cancelled', 'completed']
+  }
+})
+
+module.exports = Order
+
